# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  common.ResponseError:
    properties:
      error:
        type: string
    type: object
...
  common.ResponseNormal:
    properties:
      data: {}
      email:
        type: string
      message:
        type: string
      paging: {}
      result:
        type: boolean
    type: object
...
  dtos.CreateLevelRequest:
    properties:
      description:
        type: string
      name:
        type: string
      point_require:
        type: integer
      program_id:
        type: string
    required:
    - name
    - point_require
    - program_id
    type: object
...
  models.Expert:
    properties:
      account_id:
        type: string
      avatar_url:
        type: string
      date_of_birth:
        type: string
      expert_id:
        type: string
      expert_type:
        type: string
      full_name:
        type: string
      gender:
        type: boolean
      telephone_number:
        type: string
    required:
    - date_of_birth
    - expert_type
    - full_name
    type: object
...
host: 127.0.0.1:9000
info:
  contact: {}
  description: List APIs of Healthy Management Service
  termsOfService: http://swagger.io/terms/
  title: Healthy Service API Document
  version: "1.0"
paths:
  /admin/experts:
    get:
      consumes:
      - application/json
      description: Returns a list of experts available in the system
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Page number (default is 1)
        in: query
        name: page
        type: integer
      - description: Number of experts per page (default is 10)
        in: query
        name: limit
        type: integer
      - description: Filter is deleted (default is false)
        in: query
        name: status
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of expert
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseNormal'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Expert'
                  type: array
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/common.ResponseError'
        "401":
          description: Invalid token
          schema:
            $ref: '#/definitions/common.ResponseError'
        "403":
          description: You do not have permission to access this resource
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Retrieve a list of experts
      tags:
      - Expert
    post:
      consumes:
      - multipart/form-data
      description: Create expert profile with file image and json expert data
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Expert image file (max 10MB)
        in: formData
        name: image
        type: file
      - description: Expert data in JSON format
        in: formData
        name: metadata
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Expert created successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseNormal'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create a new expert
      tags:
      - Expert
  /admin/experts/{id}:
    delete:
      description: Soft delete an expert profile by ID. Optionally deactivate the
        linked user account.
      parameters:
      - description: Expert ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "204":
          description: No Content - Successfully deleted
        "400":
          description: Bad Request - Invalid expert ID
          schema:
            $ref: '#/definitions/common.ResponseError'
        "401":
          description: Unauthorized - Missing or invalid API key
          schema:
            $ref: '#/definitions/common.ResponseError'
        "403":
          description: Forbidden - User does not have permission
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Delete an expert profile (soft delete)
      tags:
      - Expert
    put:
      consumes:
      - multipart/form-data
      description: Updates an expert's details based on their ID.
      parameters:
      - description: Expert ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Expert profile picture
        in: formData
        name: image
        type: file
      - description: Expert information in JSON format
        in: formData
        name: metadata
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Expert updated successfully
          schema:
            $ref: '#/definitions/common.ResponseNormal'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: Internal server error during expert update
          schema:
            $ref: '#/definitions/common.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update expert information
      tags:
      - Expert
...
  /expert/levels:
    post:
      consumes:
      - application/json
      description: Creates a new level for a specified program owned by the expert.
        Requires valid JWT token.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Level details to create
        in: body
        name: level
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateLevelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Level created successfully
          schema:
            $ref: '#/definitions/common.ResponseNormal'
        "400":
          description: invalid level data or invalid program ID
          schema:
            $ref: '#/definitions/common.ResponseError'
        "401":
          description: token expired or invalid
          schema:
            $ref: '#/definitions/common.ResponseError'
        "404":
          description: program not found
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: failed to create level
          schema:
            $ref: '#/definitions/common.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create a new level for a program
      tags:
      - Level
...
schemes:
- http
- https
swagger: "2.0"